Adding a simple "base" role.
The role will add some base utility packages.

The role also copies a hosts file to /etc/hosts on each VMs.
The use of a prebaked hosts file removes the need for a working DNS server.
The hosts file only has entries for 4 servers, but can be extended very easily.