A few changes
Changed all admin scripts to use the "service" command instead of the
new "systemctl" command since we are still working with init scripts.

Disabling Zookeeper from starting at system boot.

Commenting out a line in hadoop-env.sh that was causing the init scripts
for hdfs to fail when stopping (BIGTOP-1129).